+++ BitmovinYoSpaceModule/Classes/BitmovinYospacePlayer.swift
@@ -396,7 +396,7 @@ extension BitmovinYospacePlayer: YSAnalyticObserver {
             BitLog.d("Pausing player")
             super.pause()
 
-            let adBreakPosition: AdBreakPosition = activeAdBreak?.relativeStart == 0 ? .preroll : .midroll
+            let adBreakPosition: YospaceAdBreakPosition = activeAdBreak?.relativeStart == 0 ? .preroll : .midroll
             renderer.renderTruexAd(advert: advert, adBreakPosition: adBreakPosition)
         }
         #endif
@@ -1028,5 +1028,75 @@ extension BitmovinYospacePlayer: PlayerListener {
         for listener: PlayerListener in listeners {
             listener.onDrmDataParsed?(event)
         }
+    }
+}
+
+// MARK: AdBreak Transformation
+
+extension YSAdBreak {
+    func toYospaceAdBreak(absoluteStart: Double, relativeStart: Double) -> YospaceAdBreak {
+        let yospaceAdBreak = YospaceAdBreak(
+            identifier: adBreakIdentifier(),
+            absoluteStart: absoluteStart,
+            absoluteEnd: absoluteStart + adBreakDuration(),
+            duration: adBreakDuration(),
+            relativeStart: relativeStart,
+            scheduleTime: 0,
+            replaceContentDuration: 0,
+            position: adBreakPosition().toYospaceAdBreakPosition()
+        )
+
+        // Add adverts to ad break
+        var adAbsoluteStart = absoluteStart
+        for case let advert as YSAdvert in adverts() {
+            yospaceAdBreak.register(advert.toYospaceAd(absoluteStart: adAbsoluteStart, relativeStart: relativeStart))
+            adAbsoluteStart += advert.advertDuration()
+        }
+
+        return yospaceAdBreak
+    }
+}
+
+extension YSEAdBreakPosition {
+    func toYospaceAdBreakPosition() -> YospaceAdBreakPosition {
+        switch self {
+        case .prerollPosition:
+            return .preroll
+        case .midrollPosition:
+            return .midroll
+        case .postrollPosition:
+            return .postroll
+        case .unknownPosition, _:
+            return .unknown
+        }
+    }
+}
+
+// MARK: - Ad Transformation
+
+extension YSAdvert {
+    func toYospaceAd(absoluteStart: Double, relativeStart: Double) -> YospaceAd {
+        return YospaceAd(
+            identifier: advertIdentifier(),
+            creativeId: linearCreativeElement().linearIdentifier(),
+            sequence: advertProperty("sequence"),
+            absoluteStart: absoluteStart,
+            relativeStart: relativeStart,
+            duration: advertDuration(),
+            absoluteEnd: absoluteStart + advertDuration(),
+            system: advertProperty("AdSystem"),
+            title: advertProperty("AdTitle"),
+            advertiser: advertProperty("Advertiser"),
+            hasInteractiveUnit: hasLinearInteractiveUnit(),
+            lineage: advertLineage(),
+            extensions:
+            // advertExtensions() returns the "extensions" node itself
+            // This creates a list of child "extension" nodes to be consistent with Android
+            advertExtensions()?.children().compactMap { $0 as? YSXmlNode } ?? [YSXmlNode](),
+            isFiller: isFiller(),
+            isLinear: !hasLinearInteractiveUnit(),
+            clickThroughUrl: linearCreativeElement().linearClickthroughURL(),
+            mediaFileUrl: linearCreativeElement().interactiveUnit()?.unitSource()
+        )
     }
 }